Fix "not up-to-date" errors in format script
When checking if source files needs to be reformatted, the clang-format
update script builds targets with `--check_up_to_date` and greps the
output for out of date marker files (ending with `.clang_format`). These
appear with `ERROR` in the output.

This commit builds the clang-format binary so that it is not out of
date.